refactor: introduce barrel constants.ts files
Add a local `constants.ts` file to each package (that uses constants)
so that each package can import constants from a local file. Currently,
these `constants.ts` files are all barrel files that just import the
constants from package constants and export it again. This makes it
easier to change constants as there's always just one file per package
to change.
