+++ src/clients/constants.ts
@@ -1,0 +1,5 @@
+import { INPUT_NAME_REPO_TOKEN } from "../constants";
+
+export {
+  INPUT_NAME_REPO_TOKEN,
+};
