+++ src/inputs/constants.ts
@@ -1,0 +1,27 @@
+import {
+  DEFAULT_IGNORE_GLOBS,
+  DEFAULT_IS_DRY_RUN,
+  DEFAULT_IS_STRICT_MODE,
+  DEFAULT_SVGO_V1_CONFIG_PATH,
+  DEFAULT_SVGO_V2_CONFIG_PATH,
+  DEFAULT_SVGO_VERSION,
+  INPUT_NAME_IGNORE,
+  INPUT_NAME_DRY_RUN,
+  INPUT_NAME_STRICT,
+  INPUT_NAME_SVGO_CONFIG,
+  INPUT_NAME_SVGO_VERSION,
+} from "../constants";
+
+export {
+  DEFAULT_IGNORE_GLOBS,
+  DEFAULT_IS_DRY_RUN,
+  DEFAULT_IS_STRICT_MODE,
+  DEFAULT_SVGO_V1_CONFIG_PATH,
+  DEFAULT_SVGO_V2_CONFIG_PATH,
+  DEFAULT_SVGO_VERSION,
+  INPUT_NAME_IGNORE,
+  INPUT_NAME_DRY_RUN,
+  INPUT_NAME_STRICT,
+  INPUT_NAME_SVGO_CONFIG,
+  INPUT_NAME_SVGO_VERSION,
+};
