+++ src/outputs/constants.ts
@@ -1,0 +1,15 @@
+import {
+  EVENT_PULL_REQUEST,
+  EVENT_PUSH,
+  EVENT_REPOSITORY_DISPATCH,
+  EVENT_SCHEDULE,
+  EVENT_WORKFLOW_DISPATCH,
+} from "../constants";
+
+export {
+  EVENT_PULL_REQUEST,
+  EVENT_PUSH,
+  EVENT_REPOSITORY_DISPATCH,
+  EVENT_SCHEDULE,
+  EVENT_WORKFLOW_DISPATCH,
+};
